Refactoring of all datatype conversion into new file
