+++ sciopy/com_util.py
@@ -13,7 +13,7 @@ import numpy as np
 import struct
 import sys
 from glob import glob
-
+from .datatype_conversion import *
 
 def available_serial_ports() -> list:
     """
@@ -66,27 +66,6 @@ def clTbt_dp(val: float) -> list:
     return [int(ele) for ele in struct.pack(">d", val)]
 
 
-def del_hex_in_list(lst: list) -> np.ndarray:
-    """
-    Delete the hexadecimal 0x python notation.
-
-    Parameters
-    ----------
-    lst : list
-        list of hexadecimals
-
-    Returns
-    -------
-    np.ndarray
-        cleared message
-    """
-    return np.array(
-        [
-            "0" + ele.replace("0x", "") if len(ele) == 1 else ele.replace("0x", "")
-            for ele in lst
-        ]
-    )
-
 
 def reshape_full_message_in_bursts(lst: list, ssms: EitMeasurementSetup) -> np.ndarray:
     """
@@ -130,173 +109,6 @@ def reshape_full_message_in_bursts(lst: list, ssms: EitMeasurementSetup) -> np.n
         split_list.append(lst[split * split_length: (split + 1) * split_length])
     return np.array(split_list)
 
-
-def single_hex_to_int(str_num: str) -> int:
-    """
-    Delete the hexadecimal 0x python notation.
-
-    Parameters
-    ----------
-    str_num : str
-        single hexadecimal string
-
-    Returns
-    -------
-    int
-        integer number
-    """
-    if len(str_num) == 1:
-        str_num = f"0x0{str_num}"
-    else:
-        str_num = f"0x{str_num}"
-    return int(str_num, 16)
-
-
-def bytesarray_to_float(bytes_array: np.ndarray) -> float:
-    """
-    Converts a bytes array to a float number.
-
-    Parameters
-    ----------
-    bytes_array : np.ndarray
-        array of bytes
-
-    Returns
-    -------
-    float
-        single precision float
-    """
-    bytes_array = [int(b, 16) for b in bytes_array]
-    bytes_array = bytes(bytes_array)
-    return struct.unpack("!f", bytes(bytes_array))[0]
-
-
-
-def byteintarray_to_float(bytes_array: np.ndarray) -> float:
-    """
-    Converts a bytes array to a float number. Array is array of integers representing bytes.
-
-    Parameters
-    ----------
-    bytes_array : np.ndarray
-        array of integers former being bytes
-
-    Returns
-    -------
-    float
-        single precision float
-    """
-    return struct.unpack("!f", bytes(bytes_array))[0]
-
-
-def bytesarray_to_double(bytes_array: np.ndarray) -> float:
-    """
-    Converts a bytes array to a float number.
-
-    Parameters
-    ----------
-    bytes_array : np.ndarray
-        array of bytes
-
-    Returns
-    -------
-    float
-        double precision float
-    """
-    bytes_array = [int(b, 16) for b in bytes_array]
-    bytes_array = bytes(bytes_array)
-    return struct.unpack("!d", bytes(bytes_array))[0]
-
-
-def bytesarray_to_byteslist(bytes_array: np.ndarray) -> list:
-    """
-    Converts a bytes array to a list of bytes.
-
-    Parameters
-    ----------
-    bytes_array : np.ndarray
-        array of bytes
-
-    Returns
-    -------
-    list
-        list of bytes
-    """
-    bytes_array = [int(b, 16) for b in bytes_array]
-    return bytes(bytes_array)
-
-
-def bytesarray_to_int(bytes_array: np.ndarray) -> int:
-    """
-    Converts a bytes array to int number.
-
-    Parameters
-    ----------
-    bytes_array : np.ndarray
-        array of bytes
-
-    Returns
-    -------
-    int
-        integer number
-    """
-    bytes_array = bytesarray_to_byteslist(bytes_array)
-    return int.from_bytes(bytes_array, "big")
-
-
-TWOPOWER24 = 16777216
-TWOPOWER16 = 65536
-TWOPOWER8 = 256
-
-
-def four_byte_to_int(bytelist):
-    """
-    Converts a list of 4 integers representing bytes to int.
-
-    Parameters
-    ----------
-    bytelist : np.ndarray/list of integers representing bytes MSB first
-
-    Returns
-    -------
-    int
-        integer number
-    """
-    return TWOPOWER24 * bytelist[0] + TWOPOWER16 * bytelist[1] + TWOPOWER8 * bytelist[2] + bytelist[3]
-
-
-def two_byte_to_int(bytelist):
-    """
-    Converts a list of 2 integers representing bytes to int.
-
-    Parameters
-    ----------
-    bytelist : np.ndarray/list of integers representing bytes MSB first
-
-    Returns
-    -------
-    int
-        integer number
-    """
-    return TWOPOWER8 * bytelist[0] + bytelist[1]
-
-
-def bytelist_to_int(bytelist):
-    """
-    Converts a list of integers representing bytes MSB first to int.
-    Parameters
-    ----------
-    bytelist : np.ndarray/list of integers representing bytes MSB first
-
-    Returns
-    -------
-    int
-        integer number
-    """
-    r = bytelist[-1]
-    for j in range(2, len(bytelist)):
-        r += bytelist[-j] * 2 ** ((j - 1) * 8)
-    return r
 
 
 def parse_single_frame(lst_ele: np.ndarray) -> SingleFrame:
